+++ FlixVerse/Data/Base/BaseRepository.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq.Expressions;
-using FlixVerse.Services.Database;
+using FlixVerse.Data.Context;
 using Microsoft.EntityFrameworkCore;
 
 namespace FlixVerse.Data.Base;
