+++ FlixVerse/Models/User/LoginRequest.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace FlixVerse.Models.User;
+
+public class LoginRequest
+{
+    [Required]
+    public string UsernameOrEmail { get; set; }
+    [Required]
+    public string Password { get; set; }
+}
