+++ FlixVerse/Models/User/User.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore;
 
-namespace FlixVerse.Models;
+namespace FlixVerse.Models.User;
 
 [Table("Users")]
 [Index(nameof(Username), nameof(Email), IsUnique = true)]
